Initial component restructure
Introduced a new component structure with elements, modules and
collections contained within an overall components folder inside /app
