+++ app/components/modules/uiBasics/uiBasics.scss
